move  useEffect to app)
